# openapi.yaml
openapi: 3.0.3
info:
  title: Blockchain Explorer API
  description: |
    A production-grade Ethereum blockchain indexer and query platform.

    This API provides access to indexed blockchain data including blocks, transactions,
    addresses, event logs, and real-time WebSocket streaming.

    ## Features
    - REST API for querying blockchain data
    - WebSocket streaming for real-time updates
    - Pagination support for all list endpoints
    - Health checks and metrics

    ## Base URL
    `http://localhost:8080`

    ## Authentication
    Currently, the API is public and does not require authentication.
  version: 1.0.0
  contact:
    name: API Support
    url: https://github.com/hieutt50/go-blockchain-explorer
  license:
    name: MIT
    url: https://opensource.org/licenses/MIT

servers:
  - url: http://localhost:8080
    description: Local development server
  - url: https://api.blockchain-explorer.example.com
    description: Production server (example)

tags:
  - name: Health
    description: Service health and status
  - name: Blocks
    description: Block queries and listings
  - name: Transactions
    description: Transaction queries
  - name: Addresses
    description: Address transaction history
  - name: Logs
    description: Smart contract event logs
  - name: Stats
    description: Chain statistics
  - name: Metrics
    description: Prometheus metrics

paths:
  /health:
    get:
      tags:
        - Health
      summary: Health check
      description: Check the API server and database health status
      operationId: getHealth
      responses:
        '200':
          description: Service is healthy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'
        '503':
          description: Service is unhealthy
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/HealthResponse'

  /v1/blocks:
    get:
      tags:
        - Blocks
      summary: List recent blocks
      description: Get a paginated list of recent blocks, ordered by height descending
      operationId: listBlocks
      parameters:
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/OffsetParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/BlocksResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'

  /v1/blocks/{heightOrHash}:
    get:
      tags:
        - Blocks
      summary: Get block by height or hash
      description: Get detailed information about a specific block by height or hash
      operationId: getBlock
      parameters:
        - name: heightOrHash
          in: path
          required: true
          description: Block height (e.g., 18500000) or block hash (0x + 64 hex chars)
          schema:
            type: string
          examples:
            byHeight:
              value: "18500000"
              summary: Get block by height
            byHash:
              value: "0x1234567890abcdef1234567890abcdef1234567890abcdef1234567890abcdef"
              summary: Get block by hash
      responses:
        '200':
          description: Block found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Block'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

  /v1/txs/{hash}:
    get:
      tags:
        - Transactions
      summary: Get transaction by hash
      description: Get detailed information about a specific transaction
      operationId: getTransaction
      parameters:
        - name: hash
          in: path
          required: true
          description: Transaction hash (0x + 64 hex characters)
          schema:
            type: string
            pattern: '^0x[0-9a-fA-F]{64}$'
          example: "0xabcdef1234567890abcdef1234567890abcdef1234567890abcdef1234567890"
      responses:
        '200':
          description: Transaction found
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/Transaction'
        '400':
          $ref: '#/components/responses/BadRequest'
        '404':
          $ref: '#/components/responses/NotFound'
        '500':
          $ref: '#/components/responses/InternalError'

  /v1/address/{addr}/txs:
    get:
      tags:
        - Addresses
      summary: Get address transaction history
      description: Get all transactions involving a specific address (as sender or receiver)
      operationId: getAddressTransactions
      parameters:
        - name: addr
          in: path
          required: true
          description: Ethereum address (0x + 40 hex characters)
          schema:
            type: string
            pattern: '^0x[0-9a-fA-F]{40}$'
          example: "0x742d35Cc6634C0532925a3b844Bc9e7595f0bEb0"
        - $ref: '#/components/parameters/LimitParam'
        - $ref: '#/components/parameters/OffsetParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/AddressTransactionsResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'

  /v1/logs:
    get:
      tags:
        - Logs
      summary: Query event logs
      description: Query smart contract event logs with optional filters
      operationId: queryLogs
      parameters:
        - name: address
          in: query
          required: false
          description: Contract address to filter by
          schema:
            type: string
            pattern: '^0x[0-9a-fA-F]{40}$'
          example: "0x742d35Cc6634C0532925a3b844Bc9e7595f0bEb0"
        - name: topic0
          in: query
          required: false
          description: Event signature hash (topic0)
          schema:
            type: string
            pattern: '^0x[0-9a-fA-F]{64}$'
          example: "0xddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef"
        - name: limit
          in: query
          required: false
          description: Number of logs to return (max 1000)
          schema:
            type: integer
            default: 100
            minimum: 1
            maximum: 1000
        - $ref: '#/components/parameters/OffsetParam'
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/LogsResponse'
        '400':
          $ref: '#/components/responses/BadRequest'
        '500':
          $ref: '#/components/responses/InternalError'

  /v1/stats/chain:
    get:
      tags:
        - Stats
      summary: Get chain statistics
      description: Get overall blockchain indexing statistics
      operationId: getChainStats
      responses:
        '200':
          description: Successful response
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ChainStats'
        '500':
          $ref: '#/components/responses/InternalError'

  /metrics:
    get:
      tags:
        - Metrics
      summary: Prometheus metrics
      description: Get Prometheus-formatted metrics for monitoring
      operationId: getMetrics
      responses:
        '200':
          description: Successful response
          content:
            text/plain:
              schema:
                type: string
                example: |
                  # HELP blockchain_api_requests_total Total number of API requests
                  # TYPE blockchain_api_requests_total counter
                  blockchain_api_requests_total{method="GET",endpoint="/v1/blocks"} 1234

components:
  parameters:
    LimitParam:
      name: limit
      in: query
      required: false
      description: Number of results to return
      schema:
        type: integer
        default: 25
        minimum: 1
        maximum: 100

    OffsetParam:
      name: offset
      in: query
      required: false
      description: Number of results to skip
      schema:
        type: integer
        default: 0
        minimum: 0

  schemas:
    HealthResponse:
      type: object
      properties:
        status:
          type: string
          enum: [ok, unhealthy]
          example: ok
        database:
          type: string
          enum: [healthy, unhealthy]
          example: healthy
        uptime_seconds:
          type: integer
          example: 3600
        timestamp:
          type: string
          format: date-time
          example: "2025-10-31T10:00:00Z"

    Block:
      type: object
      properties:
        height:
          type: integer
          format: int64
          example: 18500000
        hash:
          type: string
          example: "0x1234567890abcdef1234567890abcdef1234567890abcdef1234567890abcdef"
        parent_hash:
          type: string
          example: "0xabcdef1234567890abcdef1234567890abcdef1234567890abcdef1234567890"
        miner:
          type: string
          example: "0x742d35Cc6634C0532925a3b844Bc9e7595f0bEb0"
        gas_used:
          type: string
          example: "15000000"
        gas_limit:
          type: string
          example: "30000000"
        timestamp:
          type: integer
          format: int64
          example: 1698768000
        tx_count:
          type: integer
          example: 150
        orphaned:
          type: boolean
          example: false
        created_at:
          type: string
          format: date-time
          example: "2025-10-31T10:00:00Z"

    BlocksResponse:
      type: object
      properties:
        blocks:
          type: array
          items:
            $ref: '#/components/schemas/Block'
        total:
          type: integer
          example: 18500000
        limit:
          type: integer
          example: 25
        offset:
          type: integer
          example: 0

    Transaction:
      type: object
      properties:
        hash:
          type: string
          example: "0xabcdef1234567890abcdef1234567890abcdef1234567890abcdef1234567890"
        block_height:
          type: integer
          format: int64
          example: 18500000
        tx_index:
          type: integer
          example: 42
        from_addr:
          type: string
          example: "0x742d35Cc6634C0532925a3b844Bc9e7595f0bEb0"
        to_addr:
          type: string
          nullable: true
          example: "0x1234567890abcdef1234567890abcdef12345678"
        value_wei:
          type: string
          example: "1000000000000000000"
        fee_wei:
          type: string
          example: "21000000000000"
        gas_used:
          type: string
          example: "21000"
        gas_price:
          type: string
          example: "1000000000"
        nonce:
          type: integer
          format: int64
          example: 10
        success:
          type: boolean
          example: true
        created_at:
          type: string
          format: date-time
          example: "2025-10-31T10:00:00Z"

    AddressTransactionsResponse:
      type: object
      properties:
        address:
          type: string
          example: "0x742d35Cc6634C0532925a3b844Bc9e7595f0bEb0"
        transactions:
          type: array
          items:
            $ref: '#/components/schemas/Transaction'
        total:
          type: integer
          example: 1234
        limit:
          type: integer
          example: 50
        offset:
          type: integer
          example: 0

    Log:
      type: object
      properties:
        id:
          type: integer
          format: int64
          example: 12345
        tx_hash:
          type: string
          example: "0xabcdef1234567890abcdef1234567890abcdef1234567890abcdef1234567890"
        log_index:
          type: integer
          example: 0
        address:
          type: string
          example: "0x742d35Cc6634C0532925a3b844Bc9e7595f0bEb0"
        topic0:
          type: string
          nullable: true
          example: "0xddf252ad1be2c89b69c2b068fc378daa952ba7f163c4a11628f55a4df523b3ef"
        topic1:
          type: string
          nullable: true
          example: "0x000000000000000000000000742d35cc6634c0532925a3b844bc9e7595f0beb0"
        topic2:
          type: string
          nullable: true
          example: "0x0000000000000000000000001234567890abcdef1234567890abcdef12345678"
        topic3:
          type: string
          nullable: true
        data:
          type: string
          example: "0x0000000000000000000000000000000000000000000000000de0b6b3a7640000"
        created_at:
          type: string
          format: date-time
          example: "2025-10-31T10:00:00Z"

    LogsResponse:
      type: object
      properties:
        logs:
          type: array
          items:
            $ref: '#/components/schemas/Log'
        total:
          type: integer
          example: 5678
        limit:
          type: integer
          example: 100
        offset:
          type: integer
          example: 0

    ChainStats:
      type: object
      properties:
        latest_block:
          type: integer
          format: int64
          example: 18500000
        total_blocks:
          type: integer
          format: int64
          example: 18500000
        total_transactions:
          type: integer
          format: int64
          example: 125430000
        indexer_lag_blocks:
          type: integer
          example: 0
        indexer_lag_seconds:
          type: integer
          example: 2
        last_updated:
          type: string
          format: date-time
          example: "2025-10-31T10:00:00Z"

    Error:
      type: object
      properties:
        error:
          type: string
          example: "Invalid parameter"
        code:
          type: string
          example: "BAD_REQUEST"
        timestamp:
          type: string
          format: date-time
          example: "2025-10-31T10:00:00Z"

  responses:
    BadRequest:
      description: Bad request - Invalid parameters
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Invalid parameter format"
            code: "BAD_REQUEST"
            timestamp: "2025-10-31T10:00:00Z"

    NotFound:
      description: Resource not found
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Block not found"
            code: "NOT_FOUND"
            timestamp: "2025-10-31T10:00:00Z"

    InternalError:
      description: Internal server error
      content:
        application/json:
          schema:
            $ref: '#/components/schemas/Error'
          example:
            error: "Internal server error"
            code: "INTERNAL_ERROR"
            timestamp: "2025-10-31T10:00:00Z"